subsys: net: ip: random mac address
added net mac module to make sure we have one randomly generated mac
address in the system.

function_ecm and netusb uses the module to make sure the mac addresses
are random but only within 1 of each other. This is to make sure the
host end of ecm can find netusb by subtracting 1.

Signed-off-by: Nicolai Glud <nicolai.glud@prevas.dk>